bot/requirements.txt
discord-py==1.5.0
praw==7.1.0
google-cloud-datastore==1.15.1
git+git://github.com/Rapptz/discord-ext-menus
